+++ backend/canisters/user_index/api/src/updates/register_bot.rs
@@ -1,11 +1,10 @@
-use candid::{CandidType, Principal};
-use human_readable::{HumanReadablePrincipal, ToHumanReadable};
+use candid::Principal;
 use serde::{Deserialize, Serialize};
 use ts_export::ts_export;
-use types::{BotDefinition, UserId};
+use types::{BotDefinition, BotInstallationLocation, UserId};
 
 #[ts_export(user_index, register_bot)]
-#[derive(CandidType, Serialize, Deserialize, Debug)]
+#[derive(Serialize, Deserialize, Debug)]
 pub struct Args {
     pub principal: Principal,
     pub owner: UserId,
@@ -13,35 +12,11 @@ pub struct Args {
     pub avatar: Option<String>, // Image as a data URL
     pub endpoint: String,
     pub definition: BotDefinition,
+    pub permitted_install_location: Option<BotInstallationLocation>,
 }
 
 #[ts_export(user_index, register_bot)]
-#[derive(CandidType, Serialize, Deserialize, Debug)]
+#[derive(Serialize, Deserialize, Debug)]
 pub enum Response {
     Success,
-}
-
-#[derive(Serialize)]
-pub struct HumanReadableArgs {
-    principal: HumanReadablePrincipal,
-    owner: HumanReadablePrincipal,
-    name: String,
-    endpoint: String,
-    description: String,
-    definition: BotDefinition,
-}
-
-impl ToHumanReadable for Args {
-    type Target = HumanReadableArgs;
-
-    fn to_human_readable(&self) -> Self::Target {
-        HumanReadableArgs {
-            principal: self.principal.into(),
-            owner: Principal::from(self.owner).into(),
-            name: self.name.clone(),
-            endpoint: self.endpoint.clone(),
-            description: self.endpoint.clone(),
-            definition: self.definition.clone(),
-        }
-    }
 }