+++ backend/integration_tests/src/bot_tests.rs
@@ -81,6 +81,7 @@ fn e2e_command_bot_test() {
         command: BotCommandInitial {
             name: command_name.clone(),
             args: Vec::new(),
+            meta: None,
         },
         scope: BotActionScope::Chat(BotActionChatDetails {
             chat,
@@ -413,6 +414,7 @@ fn send_api_key_test(chat_type: ChatType) {
         command: BotCommandInitial {
             name: "sync_api_key".to_string(),
             args: Vec::new(),
+            meta: None,
         },
         scope: BotActionScope::Chat(BotActionChatDetails {
             chat,
@@ -619,6 +621,7 @@ fn send_multiple_updates_to_same_message(chat_type: ChatType) {
         command: BotCommandInitial {
             name: command_name.clone(),
             args: Vec::new(),
+            meta: None,
         },
         scope: BotActionScope::Chat(BotActionChatDetails {
             chat,